✨ Added endpoints to like and unlike posts and comments, comment on a post and delete posts and comments

+++ server/src/routes/userRoutes.js
@@ -29,6 +29,11 @@ router.post("/post", async (req, res) => {
     if (!authorization) {
         return res.status(401).send({ error: "You must be logged in" });
     }
+    if (!_title || !_content || (!_title && !_content)) {
+        return res
+            .status(401)
+            .send({ error: "Provide all the required fields" });
+    }
     const token = authorization.replace("Bearer ", "");
     try {
         const decoded = jwt.verify(token, "MY_SECRET_KEY");
@@ -47,11 +52,14 @@ router.post("/post", async (req, res) => {
     }
 });
 
-router.delete("/delete", async (req, res) => {
+router.delete("/post/delete", async (req, res) => {
     const { id: _id } = req.query;
     const { authorization } = req.headers;
     if (!authorization) {
         return res.status(401).send({ error: "You must be logged in" });
+    }
+    if (!_id) {
+        return res.status(401).send({ error: "Provide a post ID" });
     }
     try {
         await prisma.post.delete({
@@ -62,6 +70,27 @@ router.delete("/delete", async (req, res) => {
         res.status(200).send(`Post ${_id} deleted`);
     } catch (err) {
         res.status(422).send({ error: err.meta.cause });
+    }
+});
+
+router.delete("/comment/delete", async (req, res) => {
+    const { id: _id } = req.query;
+    const { authorization } = req.headers;
+    if (!authorization) {
+        return res.status(401).send({ error: "You must be logged in" });
+    }
+    if (!_id) {
+        return res.status(401).send({ error: "Provide a comment ID" });
+    }
+    try {
+        await prisma.comment.delete({
+            where: {
+                id: _id,
+            },
+        });
+        res.status(200).send(`Comment ${_id} deleted`);
+    } catch (err) {
+        res.status(422).send({ error: err });
     }
 });
 
@@ -81,7 +110,113 @@ router.get("/getUserPosts", async (req, res) => {
         });
         res.status(200).send(posts);
     } catch (err) {
+        res.status(422).send({ error: err });
+    }
+});
+
+router.post("/post/like", async (req, res) => {
+    const { authorization } = req.headers;
+    const { postId: _postId } = req.body;
+    if (!authorization) {
+        return res.status(401).send({ error: "You must be logged in" });
+    }
+    if (!_postId) {
+        return res.status(401).send({ error: "Provide a post ID" });
+    }
+    const token = authorization.replace("Bearer ", "");
+    try {
+        const decoded = jwt.verify(token, "MY_SECRET_KEY");
+        const userId = decoded.userId;
+        const check = await prisma.like.findMany({
+            where: {
+                userId: userId,
+                postId: _postId,
+            },
+        });
+        if (check.length !== 0) {
+            await prisma.like.delete({
+                where: {
+                    id: check[0].id,
+                },
+            });
+            return res.send("Post unliked");
+        }
+        await prisma.like.create({
+            data: {
+                userId: userId,
+                postId: _postId,
+            },
+        });
+        res.send(`Post ${_postId} liked`);
+    } catch (err) {
+        res.status(422).send({ error: err });
+    }
+});
+router.post("/comment/like", async (req, res) => {
+    const { authorization } = req.headers;
+    const { commentId: _commentId } = req.body;
+    if (!authorization) {
+        return res.status(401).send({ error: "You must be logged in" });
+    }
+    if (!_commentId) {
+        return res.status(401).send({ error: "Provide a comment ID" });
+    }
+    const token = authorization.replace("Bearer ", "");
+    try {
+        const decoded = jwt.verify(token, "MY_SECRET_KEY");
+        const userId = decoded.userId;
+        const check = await prisma.like.findMany({
+            where: {
+                userId: userId,
+                commentId: _commentId,
+            },
+        });
+        if (check.length !== 0) {
+            await prisma.like.delete({
+                where: {
+                    id: check[0].id,
+                },
+            });
+            return res.send("Comment unliked");
+        }
+        await prisma.like.create({
+            data: {
+                userId: userId,
+                commentId: _commentId,
+            },
+        });
+        res.send(`Comment ${_commentId} liked`);
+    } catch (err) {
         console.log(err);
+        res.status(422).send({ error: err });
+    }
+});
+
+router.post("/post/comment", async (req, res) => {
+    const { authorization } = req.headers;
+    const { postId: _postId, content: _content } = req.body;
+    if (!authorization) {
+        return res.status(401).send({ error: "You must be logged in" });
+    }
+    if (!_postId || !_content || (!_postId && !_content)) {
+        return res
+            .status(401)
+            .send({ error: "Provide all the required fields" });
+    }
+    const token = authorization.replace("Bearer ", "");
+    try {
+        const decoded = jwt.verify(token, "MY_SECRET_KEY");
+        const userId = decoded.userId;
+
+        await prisma.comment.create({
+            data: {
+                content: _content,
+                authorId: userId,
+                postId: _postId,
+            },
+        });
+        res.send(`Commented on post ${_postId} by user ${userId}`);
+    } catch (err) {
         res.status(422).send({ error: err });
     }
 });
